feat: improve notification handling
* remove notifications to NATS from API handlers for objects with no
  reconciler
* remove notifications to NATS from API handlers when object's
  Reconciled field is true

+++ internal/kubernetesruntime/kubernetes_runtime_instance.go
@@ -1,6 +1,7 @@
 package kubernetesruntime
 
 import (
+	"errors"
 	"fmt"
 	"time"
 
@@ -21,15 +22,6 @@ func kubernetesRuntimeInstanceCreated(
 	kubernetesRuntimeInstance *v0.KubernetesRuntimeInstance,
 	log *logr.Logger,
 ) error {
-	// TODO: remove this once rebased onto Randy's changes that will remove the
-	// notification from coming in if reconciled
-	// if a cluster instance is created by another mechanism and being
-	// registered in the system with Reconciled=true, there's no need to do
-	// anything - return immediately without error
-	if *kubernetesRuntimeInstance.Reconciled {
-		return nil
-	}
-
 	// get runtime definition
 	kubernetesRuntimeDefinition, err := client.GetKubernetesRuntimeDefinitionByID(
 		r.APIClient,
@@ -82,7 +74,7 @@ func kubernetesRuntimeInstanceCreated(
 	return nil
 }
 
-// kubernetesRuntimeInstanceUpdate reconciles state for a kubernetes
+// kubernetesRuntimeInstanceUpdated reconciles state for a kubernetes
 // runtime instance whenever it is changed.
 func kubernetesRuntimeInstanceUpdated(
 	r *controller.Reconciler,
@@ -119,7 +111,7 @@ func kubernetesRuntimeInstanceUpdated(
 	return nil
 }
 
-// kubernetesRuntimeInstanceCreated reconciles state for a kubernetes
+// kubernetesRuntimeInstanceDeleted reconciles state for a kubernetes
 // runtime instance whenever it is removed.
 func kubernetesRuntimeInstanceDeleted(
 	r *controller.Reconciler,
@@ -136,9 +128,12 @@ func kubernetesRuntimeInstanceDeleted(
 		return fmt.Errorf("failed to kubernetes runtime definition by ID: %w", err)
 	}
 
-	// check for workload instances - delete if any present as related
-	// infrastructure, such as cloud provider load balancers may prevent runtime
-	// instance deletion
+	// check for workload instances:
+	// return error if ForceDelete field on k8s runtime instance is false
+	// delete workload instances if ForceDelete is true
+	// workload instances must be removed prior to runtime deletion as workloads
+	// may have attached infrastructure, e.g. load balancers that will prevent
+	// infrastructure clean up
 	workloadInstances, err := client.GetWorkloadInstancesByKubernetesRuntimeInstanceID(
 		r.APIClient,
 		r.APIServer,
@@ -147,25 +142,27 @@ func kubernetesRuntimeInstanceDeleted(
 	if err != nil {
 		return fmt.Errorf("failed to get workload instances running in kubernetes runtime: %w", err)
 	}
-
 	if len(*workloadInstances) > 0 {
-		for _, wi := range *workloadInstances {
-			_, err := client.DeleteWorkloadInstance(
-				r.APIClient,
-				r.APIServer,
-				*wi.ID,
-			)
-			if err != nil {
-				return fmt.Errorf("failed to delete workload instance with ID %d: %w", wi.ID, err)
+		if *kubernetesRuntimeInstance.ForceDelete {
+			for _, wi := range *workloadInstances {
+				_, err := client.DeleteWorkloadInstance(
+					r.APIClient,
+					r.APIServer,
+					*wi.ID,
+				)
+				if err != nil {
+					return fmt.Errorf("failed to delete workload instance with ID %d: %w", wi.ID, err)
+				}
 			}
+			// TODO: wait for workload instances to be deleted before deleting runtime
+			// without a dumb sleep
+			time.Sleep(time.Second * 300)
+		} else {
+			return errors.New("kubernetes runtime instance cannot be deleted - workload instances are present and ForceDelete is false")
 		}
-
-		// TODO: wait for workload instances to be deleted before deleting runtime
-		// without a dumb wait
-		time.Sleep(time.Second * 300)
-
-		// TODO: delete support services and control plane components
 	}
+
+	// TODO: delete support services and control plane components
 
 	// delete kubernetes runtime instance
 	switch *kubernetesRuntimeDefinition.InfraProvider {
